Move Array2D class to separate file and rename DFloat
