+++ .gitignore
@@ -19,3 +19,4 @@ ext/fast_statistics/mkmf.log
 
 squiggles.txt
 README.md.html
+.tool-versions
