+++ ext/fast_statistics/debug.h
@@ -2,19 +2,20 @@
 #ifndef DEBUG_H
 #define DEBUG_H
 
-#include <x86intrin.h>
 #include <stdio.h>
 #include <string.h>
+#include <x86intrin.h>
 
 #define CONCAT_(a, b) a##b
-#define CONCAT(a,b) CONCAT_(a, b)
+#define CONCAT(a, b) CONCAT_(a, b)
 #define PROFILE DebugTimer CONCAT(Timer, __COUNTER__)(__func__, __LINE__);
 
 struct DebugTimer {
   char* name;
   unsigned long long counter;
-  
-  DebugTimer(const char* function_name, int line_number) {
+
+  DebugTimer(const char* function_name, int line_number)
+  {
     name = (char*)malloc(200 * sizeof(char));
 
     strcpy(name, function_name);
@@ -24,7 +25,8 @@ struct DebugTimer {
     counter = __rdtsc();
   }
 
-  ~DebugTimer() {
+  ~DebugTimer()
+  {
     printf("\n%30s:\t %-10llu", name, __rdtsc() - counter);
     free(name);
     fflush(stdout);
@@ -32,5 +34,5 @@ struct DebugTimer {
 };
 #endif
 #else
-#define PROFILE();
+#define PROFILE() ;
 #endif
