+++ ext/fast_statistics/utils.cpp
@@ -1,0 +1,32 @@
+#include "utils.h"
+
+VALUE
+build_results_hashes(Stats* stats, int num_variables)
+{
+  VALUE a_results = rb_ary_new();
+
+  VALUE s_min = rb_sym("min");
+  VALUE s_max = rb_sym("max");
+  VALUE s_mean = rb_sym("mean");
+  VALUE s_median = rb_sym("median");
+  VALUE s_q1 = rb_sym("q1");
+  VALUE s_q3 = rb_sym("q3");
+  VALUE s_standard_deviation = rb_sym("standard_deviation");
+
+  for (int i = 0; i < num_variables; i++) {
+    VALUE h_result = rb_hash_new();
+    Stats var_stats = stats[i];
+
+    rb_hash_aset(h_result, s_min, DBL2NUM(var_stats.min));
+    rb_hash_aset(h_result, s_max, DBL2NUM(var_stats.max));
+    rb_hash_aset(h_result, s_mean, DBL2NUM(var_stats.mean));
+    rb_hash_aset(h_result, s_median, DBL2NUM(var_stats.median));
+    rb_hash_aset(h_result, s_q1, DBL2NUM(var_stats.q1));
+    rb_hash_aset(h_result, s_q3, DBL2NUM(var_stats.q3));
+    rb_hash_aset(h_result, s_standard_deviation, DBL2NUM(var_stats.standard_deviation));
+
+    rb_ary_push(a_results, h_result);
+  }
+
+  return a_results;
+}
