+++ ext/fast_statistics/utils.h
@@ -1,0 +1,8 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include "dfloat.h"
+#include "fast_statistics.h"
+VALUE
+build_results_hashes(Stats* stats, int num_variables);
+#endif
